#261 Delete + privileged + partial orchestrator
