--- components/services/delete.ts
@@ -1,21 +1,0 @@
-import { ApiWithMessage } from '../utils/api';
-
-async function deleteInputItems({
-  body,
-}: {
-  body: {
-    itemIds: string[];
-    container: 'input_invested' | 'input_transactions';
-  };
-}) {
-  return await ApiWithMessage(
-    `/api/delete/delete_input_items`,
-    'Deleting item',
-    'Item deleted',
-    'POST',
-    body,
-    'application/json'
-  );
-}
-
-export { deleteInputItems };
