+++ components/services/orchestrator.ts
@@ -1,22 +1,4 @@
-import { ApiWithMessage, regularFetch } from '../utils/api';
-
-function startOrchestrator({
-  body,
-}: {
-  body: {
-    functionName: 'stocktracker_orchestrator';
-    daysToUpdate: number | 'all';
-  };
-}) {
-  ApiWithMessage(
-    '/api/orchestrator/start',
-    'Calling Orchestrator',
-    'Orchestration called, This will take a while',
-    'POST',
-    body,
-    'multipart/form-data'
-  );
-}
+import { regularFetch } from '../utils/api';
 
 function fetchOrchestratorList({
   body,
@@ -25,8 +7,8 @@ function fetchOrchestratorList({
   background,
 }: {
   body: { days: number | 'all' };
-  dispatcher?: any;
-  abortController?: AbortController;
+  dispatcher: any;
+  abortController: AbortController;
   background?: boolean;
 }) {
   regularFetch({
@@ -39,43 +21,4 @@ function fetchOrchestratorList({
   });
 }
 
-function purgeOrchestrator({
-  body,
-}: {
-  body: {
-    instanceId: string;
-  };
-}) {
-  ApiWithMessage(
-    '/api/orchestrator/purge',
-    'Purging orchestrator',
-    'Orchestrator purged',
-    'POST',
-    body,
-    'multipart/form-data'
-  );
-}
-
-function terminateOrchestrator({
-  body,
-}: {
-  body: {
-    instanceId: string;
-  };
-}) {
-  ApiWithMessage(
-    '/api/orchestrator/terminate',
-    'Terminating orchestrator',
-    'Orchestrator terminated',
-    'POST',
-    body,
-    'multipart/form-data'
-  );
-}
-
-export {
-  startOrchestrator,
-  fetchOrchestratorList,
-  purgeOrchestrator,
-  terminateOrchestrator,
-};
+export { fetchOrchestratorList };
